--- lib/Constants.dart
@@ -1,17 +1,0 @@
-//För drop down menyn.
-
-class Constants {
-  static const String done = 'done';
-  static const String undone = 'undone';
-  static const String all = 'all';
-
-  static const List<String> choices = <String>[
-    done,
-    undone,
-    all,
-  ];
-}
-
-class Filter {
-  static String show = 'all';
-}
