+++ pubspec.yaml
@@ -23,13 +23,11 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-    
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-  provider: ^4.3.2+2
 
 dev_dependencies:
   flutter_test:
@@ -40,6 +38,8 @@ dev_dependencies:
 
 # The following section is specific to Flutter.
 flutter:
+  assets:
+    - assets/asset2.jpg
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
